Querify clashing_extern_declarations lint.
